requirements.txt
discord.py
ffmpeg-python
PyNaCl

#pip install -r requirements.txt (easiest way via python)